delete the content folder; move search compoent to header compoent; complete the header component
