+++ src/styles/_paging.scss
@@ -8,6 +8,7 @@
   button {
     padding: 6px;
     background-color: #eeeeee;
+    outline: none;
   }
   button:hover {
     color:#eeeeee;
