+++ src/styles/index.scss
@@ -1,3 +1,4 @@
+@import './header';
 @import './content';
 @import './search';
 @import './paging';
